Extract BaseOpenSSHKeyFile for companion-public-key support
OpenSSHKeyFile and OpenSSHKeyV1KeyFile each carried an identical copy of
the companion-public-key wiring: the CompanionPublicKey field, the
getPublic()/getType() overrides that check it, and the three init()
overloads that populate it. That duplicated wiring across two classes in
two packages is what let the argument/field mix-up fixed in #1057 slip in.

Pull it up into a shared abstract base, BaseOpenSSHKeyFile, that both now
extend. Each provider is left with only what actually differs: its
readKeyPair() (openssh-key-v1 parsing vs. the PKCS8 parser strategy),
its Factory, and - for the v1 provider - setSecurityKeySigner().

BaseFileKeyProvider, PKCS8KeyFile and PuTTYKeyFile are untouched, so no
behaviour changes for the non-OpenSSH formats. (Pushing companion support
all the way into BaseFileKeyProvider, honouring the FileKeyProvider
interface which declares init(priv, pub) for every provider, is the
alternative - deliberately not done here to keep PKCS8/PuTTY unchanged.)
